scripts/train_voice.py
import os
from TTS.api import TTS
from trainer import Trainer, TrainerArgs
from TTS.tts.configs.xtts_config import XttsConfig
from TTS.tts.models.xtts import Xtts

# 1. Configuration
OUT_PATH = "/home/ubuntu/digital_twin_project/output/"
DATA_PATH = "/home/ubuntu/digital_twin_project/audio_data/"

# Define Model config (Simplified for XTTS v2)
config = XttsConfig()
config.load_json("/home/ubuntu/.local/lib/python3.10/site-packages/TTS/tts/configs/xtts_config.json") # Path varies by install

# 2. Initialize the Model
model = Xtts.init_from_config(config)

# 3. Setup Trainer
# Note: This is a placeholder. Real XTTS fine-tuning requires 
# specific GPT-formatting. For this guide, we will use the 
# CLI command in the shell script which is more robust.
print("Please use the CLI command provided in run_pipeline.sh for XTTS fine-tuning.")