requirements.txt
torch
torchaudio
openai-whisper
pydub
pandas